Move message traits to appropriate modules
